feat: 맛집 생성, 수정, 삭제에 auth 추가

+++ src/main/java/com/culturefinder/songdodongnae/delicious_spot/service/DeliciousSpotService.java
@@ -6,6 +6,10 @@ import com.culturefinder.songdodongnae.delicious_spot.domain.DeliciousSpot;
 import com.culturefinder.songdodongnae.delicious_spot.dto.DeliciousSpotReqDto;
 import com.culturefinder.songdodongnae.delicious_spot.dto.DeliciousSpotResDto;
 import com.culturefinder.songdodongnae.delicious_spot.repository.DeliciousSpotRepository;
+import com.culturefinder.songdodongnae.exception.CustomException;
+import com.culturefinder.songdodongnae.user.domain.Role;
+import com.culturefinder.songdodongnae.user.domain.User;
+import com.culturefinder.songdodongnae.user.repository.UserRepository;
 import com.culturefinder.songdodongnae.utils.CustomPage;
 import com.culturefinder.songdodongnae.s3.S3UploadService;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +17,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.*;
+
+import static com.culturefinder.songdodongnae.exception.ErrorCode.ENTITY_NOT_FOUND;
+import static com.culturefinder.songdodongnae.exception.ErrorCode.FORBIDDEN;
 
 @Service
 @Transactional
@@ -22,8 +29,14 @@ public class DeliciousSpotService {
     private final DeliciousSpotRepository deliciousSpotRepository;
     private final BookmarkRepository bookmarkRepository;
     private final S3UploadService s3UploadService;
+    private final UserRepository userRepository;
 
-    public DeliciousSpotResDto createDeliciousSpot(DeliciousSpotReqDto deliciousSpotReqDto) {
+    public DeliciousSpotResDto createDeliciousSpot(DeliciousSpotReqDto deliciousSpotReqDto, Long userId) {
+        User findUser = userRepository.findById(userId)
+                .orElseThrow(()-> new CustomException(ENTITY_NOT_FOUND));
+        if (findUser.getRole() != Role.ROLE_ADMIN) {
+            throw new CustomException(FORBIDDEN);
+        }
         DeliciousSpot deliciousSpot = deliciousSpotReqDto.toEntity();
         DeliciousSpot savedDeliciousSpot = deliciousSpotRepository.saveDeliciousSpot(deliciousSpot);
         return DeliciousSpotResDto.fromEntity(savedDeliciousSpot);
@@ -40,12 +53,22 @@ public class DeliciousSpotService {
         return DeliciousSpotResDto.fromEntity(deliciousSpot, isBookmarked);
     }
 
-    public DeliciousSpotResDto updateDeliciousSpot(Long id, DeliciousSpotReqDto deliciousSpotReqDto) {
+    public DeliciousSpotResDto updateDeliciousSpot(Long id, DeliciousSpotReqDto deliciousSpotReqDto, Long userId) {
+        User findUser = userRepository.findById(userId)
+                .orElseThrow(()-> new CustomException(ENTITY_NOT_FOUND));
+        if (findUser.getRole() != Role.ROLE_ADMIN) {
+            throw new CustomException(FORBIDDEN);
+        }
         DeliciousSpot updatedDeliciousSpot = deliciousSpotRepository.updateDeliciousSpot(id, deliciousSpotReqDto.toEntity());
         return DeliciousSpotResDto.fromEntity(updatedDeliciousSpot);
     }
 
-    public void deleteDeliciousSpot(Long id) {
+    public void deleteDeliciousSpot(Long id, Long userId) {
+        User findUser = userRepository.findById(userId)
+                .orElseThrow(()-> new CustomException(ENTITY_NOT_FOUND));
+        if (findUser.getRole() != Role.ROLE_ADMIN) {
+            throw new CustomException(FORBIDDEN);
+        }
         DeliciousSpot deliciousSpot = deliciousSpotRepository.findDeliciousSpotById(id);
         if (deliciousSpot.getThumbnailImageUrl() != null) {
             s3UploadService.deleteFile(deliciousSpot.getThumbnailImageUrl());
